refactor: 스타일링 분기 처리를 컴포넌트에서 하도록 변경
