+++ src/components/common/Button/Button.styles.tsx
@@ -1,43 +1,11 @@
 import styled, { css } from 'styled-components';
 
 export const Button = styled.button<{
-  size: 'SS' | 'S' | 'M' | 'L';
   view: 'white' | 'black';
 }>`
   padding: 0;
 
   cursor: pointer;
-
-  ${({ size }) => {
-    switch (size) {
-      case 'SS':
-        return css`
-          width: 2.4rem;
-          height: 1.4rem;
-        `;
-      case 'S':
-        return css`
-          width: 4.2rem;
-          height: 3rem;
-        `;
-      case 'M':
-        return css`
-          width: 9.8rem;
-          height: 3.5rem;
-
-          font-weight: 400;
-          font-size: 16px;
-        `;
-      case 'L':
-        return css`
-          width: 38.8rem;
-          height: 7.3rem;
-
-          font-weight: 400;
-          font-size: 24px;
-        `;
-    }
-  }}
 
   ${({ view }) => {
     switch (view) {
@@ -57,4 +25,30 @@ export const Button = styled.button<{
       }
     }
   }}
+`;
+
+export const ExtraSmallButton = styled(Button)`
+  width: 2.4rem;
+  height: 1.4rem;
+`;
+
+export const SmallButton = styled(Button)`
+  width: 4.2rem;
+  height: 3rem;
+`;
+
+export const MediumButton = styled(Button)`
+  width: 9.8rem;
+  height: 3.5rem;
+
+  font-weight: 400;
+  font-size: 16px;
+`;
+
+export const LargeButton = styled(Button)`
+  width: 38.8rem;
+  height: 7.3rem;
+
+  font-weight: 400;
+  font-size: 24px;
 `;
